+++ relay-server/src/endpoints/playstation.rs
@@ -4,7 +4,6 @@
 use axum::extract::{DefaultBodyLimit, Request};
 use axum::response::IntoResponse;
 use axum::routing::{MethodRouter, post};
-use http::StatusCode;
 use multer::{Field, Multipart};
 use relay_config::Config;
 use relay_dynamic_config::Feature;
@@ -22,7 +21,6 @@ use crate::managed::Managed;
 use crate::middlewares;
 use crate::service::ServiceState;
 use crate::services::outcome::DiscardReason;
-use crate::services::projects::cache::Project;
 use crate::services::upload::Upload;
 use crate::utils::{self, AttachmentStrategy};
 
@@ -64,9 +62,62 @@ struct Parts {
     upload: &'static [&'static str],
 }
 
-struct PlaystationAttachmentStrategy<'a> {
-    upload_service: Option<&'a Addr<Upload>>,
+struct UploadContext<'a> {
+    upload: &'a Addr<Upload>,
     scoping: Scoping,
+}
+
+/// Created an [UploadContext].
+///
+/// Requires both the `endpoint_fetch_config_enabled` option and `PlaystationUploads`
+/// feature to be enabled as well as attachments not being ratelimited.
+async fn upload_context<'a>(
+    state: &'a ServiceState,
+    meta: &RequestMeta,
+) -> Result<Option<UploadContext<'a>>, BadStoreRequest> {
+    if !state
+        .global_config_handle()
+        .current()
+        .options
+        .endpoint_fetch_config_enabled
+    {
+        return Ok(None);
+    }
+
+    let project = state
+        .project_cache_handle()
+        .ready(meta.public_key(), state.config().query_timeout())
+        .await
+        .ok_or(BadStoreRequest::ProjectUnavailable)?;
+
+    let project_config = project
+        .state()
+        .clone()
+        .enabled()
+        .ok_or(BadStoreRequest::EventRejected(DiscardReason::ProjectId))?;
+
+    let scoping = project_config
+        .scoping(meta.public_key())
+        .ok_or(BadStoreRequest::EventRejected(DiscardReason::ProjectId))?;
+
+    let attachment_rate_limits = project.rate_limits().current_limits().check_with_quotas(
+        project_config.get_quotas(),
+        scoping.item(DataCategory::Attachment),
+    );
+
+    match project_config.has_feature(Feature::PlaystationUploads)
+        && !attachment_rate_limits.is_limited()
+    {
+        true => Ok(Some(UploadContext {
+            upload: state.upload(),
+            scoping,
+        })),
+        false => Ok(None),
+    }
+}
+
+struct PlaystationAttachmentStrategy<'a> {
+    upload_context: Option<UploadContext<'a>>,
 }
 
 impl<'a> AttachmentStrategy for PlaystationAttachmentStrategy<'a> {
@@ -87,16 +138,16 @@ impl<'a> AttachmentStrategy for PlaystationAttachmentStrategy<'a> {
         item: Managed<Item>,
         config: &Config,
     ) -> Result<Option<Managed<Item>>, multer::Error> {
-        match self.upload_service {
-            Some(upload) if self.infer_type(&field) != AttachmentType::Prosperodump => {
+        match &self.upload_context {
+            Some(upload_context) if self.infer_type(&field) != AttachmentType::Prosperodump => {
                 let content_type = field.content_type().map(ToString::to_string);
                 Ok(common::upload_to_objectstore(
                     field,
                     content_type,
                     item,
                     config,
-                    self.scoping,
-                    upload,
+                    upload_context.scoping,
+                    upload_context.upload,
                     "playstation",
                 )
                 .await)
@@ -127,48 +178,33 @@ async fn multipart_to_envelope(
     multipart: Multipart<'static>,
     meta: RequestMeta,
     state: &ServiceState,
-    project: &Project<'_>,
+    upload_context: Option<UploadContext<'_>>,
 ) -> Result<Managed<Box<Envelope>>, BadStoreRequest> {
-    let project_config = project
-        .state()
-        .clone()
-        .enabled()
-        .ok_or(BadStoreRequest::EventRejected(DiscardReason::ProjectId))?;
-    let scoping = project_config
-        .scoping(meta.public_key())
-        .ok_or(BadStoreRequest::EventRejected(DiscardReason::ProjectId))?;
-    let attachment_rate_limits = || {
-        project.rate_limits().current_limits().check_with_quotas(
-            project_config.get_quotas(),
-            scoping.item(DataCategory::Attachment),
-        )
-    };
-    let upload_service = (project_config.has_feature(Feature::PlaystationUploads)
-        && !attachment_rate_limits().is_limited())
-    .then_some(state.upload());
-    let attachment_strategy = PlaystationAttachmentStrategy {
-        scoping,
-        upload_service,
-    };
     let mut items = utils::multipart_items(
         multipart,
         state.config(),
-        attachment_strategy,
+        PlaystationAttachmentStrategy { upload_context },
         &meta,
         state.outcome_aggregator(),
     )
     .await?;
 
-    let prosperodump_item = items
-        .iter_mut()
-        .find(|item| item.attachment_type() == Some(AttachmentType::Prosperodump))
-        .ok_or(BadStoreRequest::MissingProsperodump)?;
-    prosperodump_item.modify(|inner, _| inner.set_payload(OctetStream, inner.payload()));
-    validate_prosperodump(&prosperodump_item.payload())?;
+    items.try_modify(|inner, _| -> Result<(), BadStoreRequest> {
+        let prosperodump = inner
+            .iter_mut()
+            .find(|item| item.attachment_type() == Some(AttachmentType::Prosperodump))
+            .ok_or(BadStoreRequest::MissingProsperodump)?;
+        let payload = prosperodump.payload();
+        validate_prosperodump(&payload)?;
+        prosperodump.set_payload(OctetStream, payload);
+        Ok(())
+    })?;
 
     let event_id = common::event_id_from_items(&items)?.unwrap_or_else(EventId::new);
-    let envelope =
-        common::managed_items_to_envelope(items, meta, state.outcome_aggregator(), event_id);
+    let envelope = items.map(|items, records| {
+        records.modify_by(DataCategory::Error, 1);
+        Box::new(Envelope::from_request(Some(event_id), meta).with_items(items))
+    });
     Ok(envelope)
 }
 
@@ -183,15 +219,9 @@ async fn handle(
         return Ok(axum::Json(create_data_request_response()).into_response());
     }
 
-    let config = state.config();
-    let project = state
-        .project_cache_handle()
-        .ready(meta.public_key(), config.query_timeout())
-        .await
-        .ok_or(StatusCode::SERVICE_UNAVAILABLE)?;
-
+    let upload_context = upload_context(&state, &meta).await?;
     let multipart = utils::multipart_from_request(request)?;
-    let mut envelope = multipart_to_envelope(multipart, meta, &state, &project).await?;
+    let mut envelope = multipart_to_envelope(multipart, meta, &state, upload_context).await?;
     envelope.modify(|inner, _| inner.require_feature(Feature::PlaystationIngestion));
 
     let id = envelope.event_id();