+++ relay-server/src/utils/mod.rs
@@ -3,6 +3,7 @@ mod dynamic_sampling;
 mod error;
 mod multipart;
 mod param_parser;
+#[cfg(feature = "processing")]
 mod pick;
 #[cfg(all(sentry, feature = "processing"))]
 pub mod playstation;
@@ -33,6 +34,7 @@ pub use self::multipart::*;
 #[cfg(feature = "processing")]
 pub use self::native::*;
 pub use self::param_parser::*;
+#[cfg(feature = "processing")]
 pub use self::pick::*;
 pub use self::rate_limits::*;
 pub use self::retry::*;